Replace useCreatePayPal with new creator method
